sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fairlens"
version = "1.0.0"
description = "Enterprise AI bias detection SDK for ML models — detect, monitor, and remediate unfairness"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.9"
authors = [
    { name = "FairLens Team" },
]
keywords = ["fairness", "bias", "ml", "audit", "governance", "equity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fairlearn>=0.10",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "numpy>=1.24",
    "google-cloud-bigquery>=3.10",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/your-org/fairlens"
Documentation = "https://fairlens.dev"
Repository = "https://github.com/your-org/fairlens"

[tool.setuptools.packages.find]
include = ["fairlens*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
